Removed higher api level drawables and added fraction support

+++ app/src/main/java/com/assignment/payments/model/Payment.java
@@ -2,7 +2,7 @@ package com.assignment.payments.model;
 
 public class Payment {
     private final String type;
-    private final int amount;
+    private final double amount;
     private String provider;
     private String transactionReference;
 
@@ -11,7 +11,7 @@ public class Payment {
         this.amount = amount;
     }
 
-    public Payment(String type, int amount, String provider, String transactionReference) {
+    public Payment(String type, double amount, String provider, String transactionReference) {
         this.type = type;
         this.amount = amount;
         this.provider = provider;
@@ -22,7 +22,7 @@ public class Payment {
         return type;
     }
 
-    public int getAmount() {
+    public double getAmount() {
         return amount;
     }
